# internal/infrastructure/http/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/stocks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna una lista paginada en base a los filtros aplicados",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stocks"
                ],
                "summary": "Obtener stocks paginados",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Esquema JWT. Usar:'Bearer {token}'",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Búsqueda por ticker o nombre",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Ordenar resultados. Valores: tendency-asc, tendency-desc, price-asc, price-desc, ticker-asc, ticker-desc, date",
                        "name": "orderby",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Precio mínimo",
                        "name": "greater",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Precio máximo",
                        "name": "lower",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Número de página. Default: 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Tamaño de página. Default: 20",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StockViewPaginated"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/mappers.HttpErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/mappers.HttpErrResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/stocks/favorites": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna la lista de stocks asociados al usuario autenticado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stocks"
                ],
                "summary": "Obtener stocks favoritos del usuario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Esquema JWT. Usar:'Bearer {token}'",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Búsqueda por ticker o nombre",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Ordenar resultados. Valores: tendency-asc, tendency-desc, price-asc, price-desc, ticker-asc, ticker-desc, date",
                        "name": "orderby",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Precio mínimo",
                        "name": "greater",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Precio máximo",
                        "name": "lower",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Número de página. Default: 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Tamaño de página. Default: 20",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StockViewPaginated"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/mappers.HttpErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/mappers.HttpErrResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/stocks/{stockID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtener un stock por su ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stocks"
                ],
                "summary": "Obtener stock",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Esquema JWT. Usar:'Bearer {token}'",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID del stock",
                        "name": "stockID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PopulatedStockView"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/mappers.HttpErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/mappers.HttpErrResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/stocks/{stockID}/favorites": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Asocia un stock existente al usuario autenticado",
                "tags": [
                    "stocks"
                ],
                "summary": "Registrar stock al usuario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Esquema JWT. Usar:'Bearer {token}'",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID del stock",
                        "name": "stockID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mappers.HttpErrResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/mappers.HttpErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/mappers.HttpErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/mappers.HttpErrResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remueve un stock asociado al usuario autenticado",
                "tags": [
                    "stocks"
                ],
                "summary": "Eliminar stock del usuario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Esquema JWT. Usar:'Bearer {token}'",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID del stock",
                        "name": "stockID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mappers.HttpErrResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/mappers.HttpErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/mappers.HttpErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/mappers.HttpErrResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "mappers.HttpErrResponse": {
            "type": "object",
            "properties": {
                "issues": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        " \"field2\"": " \"error2\"}",
                        "{\"field1\"": " \"error1\""
                    }
                },
                "message": {
                    "type": "string",
                    "example": "mensaje generico de error"
                },
                "name": {
                    "type": "string",
                    "example": "nombre especifico del error"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
...
        "models.CompanyView": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2021-01-01T12:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "1"
                },
                "market_id": {
                    "type": "string",
                    "example": "1"
                },
                "name": {
                    "type": "string",
                    "example": "Company 1"
                }
            }
        },
        "models.MarketView": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2021-01-01T12:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "1"
                },
                "name": {
                    "type": "string",
                    "example": "Market 1"
                }
            }
        },
...
        "models.PopulatedStockView": {
            "type": "object",
            "properties": {
                "company": {
                    "$ref": "#/definitions/models.CompanyView"
                },
                "company_id": {
                    "type": "string",
                    "example": "1"
                },
                "id": {
                    "type": "string",
                    "example": "1"
                },
                "is_saved": {
                    "type": "boolean",
                    "example": true
                },
                "isin": {
                    "type": "string",
                    "example": "US0378331005"
                },
                "last_register": {
                    "$ref": "#/definitions/models.StockRegisterView"
                },
                "market": {
                    "$ref": "#/definitions/models.MarketView"
                },
                "market_id": {
                    "type": "string",
                    "example": "1"
                },
                "name": {
                    "type": "string",
                    "example": "apple inc"
                },
                "registered": {
                    "type": "string",
                    "example": "2021-01-01T12:00:00Z"
                },
                "ticker": {
                    "type": "string",
                    "example": "AAPL"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2021-01-01T12:00:00Z"
                }
            }
        },
...
        "models.StockRegisterView": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2021-01-01T12:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "1"
                },
                "price": {
                    "type": "number",
                    "example": 100
                },
                "stock_id": {
                    "type": "string",
                    "example": "1"
                },
                "tendency": {
                    "type": "string",
                    "example": "up"
                }
            }
        },
...
        "models.StockView": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "string",
                    "example": "1"
                },
                "id": {
                    "type": "string",
                    "example": "1"
                },
                "isin": {
                    "type": "string",
                    "example": "US0378331005"
                },
                "market_id": {
                    "type": "string",
                    "example": "1"
                },
                "name": {
                    "type": "string",
                    "example": "apple inc"
                },
                "registered": {
                    "type": "string",
                    "example": "2021-01-01T12:00:00Z"
                },
                "ticker": {
                    "type": "string",
                    "example": "AAPL"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2021-01-01T12:00:00Z"
                }
            }
        },
        "models.StockViewPaginated": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.StockView"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_size": {
                    "type": "integer"
                }
            }
        },
...
    }
}